chore(refactor): organize components for gas details and raw input details
