+++ src/pages/tx/[tid].tsx
@@ -1,9 +1,7 @@
 import BigNumber from "bignumber.js";
-import clsx from "clsx";
 import { Dispatch, SetStateAction, useState } from "react";
-import { FiCopy, FiChevronUp, FiChevronDown } from "react-icons/fi";
+import { FiCopy } from "react-icons/fi";
 
-import { useUnitSuffix } from "hooks/useUnitSuffix";
 import useWindowDimensions from "hooks/useWindowDimensions";
 import { formatDateToUTC, getDuration } from "shared/durationHelper";
 import { transformTransactionData } from "shared/transactionDataHelper";
@@ -23,6 +21,8 @@ import { RejectedCross } from "@components/icons/RejectedCross";
 import { GreenTickIcon } from "@components/icons/GreenTickIcon";
 import { NetworkConnection } from "@contexts/Environment";
 import BoldedTitle from "./_components/BoldedTitle";
+import GasDetails from "./_components/GasDetails";
+import RawInput from "./_components/RawInput";
 
 function Transaction({ txDetails }: { txDetails: TransactionI }) {
   const amount = { value: txDetails.value, symbol: txDetails.symbol };
@@ -66,7 +66,7 @@ function Transaction({ txDetails }: { txDetails: TransactionI }) {
             type={txDetails.type}
           />
           <div className="border-b border-black-600 lg:my-11 md:mt-9 md:mb-11 mt-10 mb-[52px]" />
-          <TransactionDetailSegmentTwo
+          <GasDetails
             gasPrice={gasPrice}
             gasLimit={txDetails.gasLimit}
             gasUsed={txDetails.gasUsed}
@@ -74,8 +74,8 @@ function Transaction({ txDetails }: { txDetails: TransactionI }) {
             from={txDetails.from}
             to={txDetails.to}
             transactionType={txDetails.transactionType}
-            hex={txDetails.rawInput}
           />
+          <RawInput hex={txDetails.rawInput} />
         </div>
       </GradientCardContainer>
     </div>
@@ -563,530 +563,6 @@ function TransactionDetailSegmentOne({
                 />
               </div>
             )}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-interface TransactionDetailSegmentTwoProps {
-  gasPrice: {
-    value: string;
-    symbol: string;
-  };
-  gasLimit: string;
-  gasUsed: string;
-  gasUsedPercentage: string;
-  from: string;
-  to: string | null;
-  transactionType: TransactionType;
-  hex: string;
-}
-
-function TransactionDetailSegmentTwo({
-  gasPrice,
-  gasLimit,
-  gasUsed,
-  gasUsedPercentage,
-  from,
-  to,
-  transactionType,
-  hex,
-}: TransactionDetailSegmentTwoProps) {
-  const [isRawInputExpanded, setIsRawInputExpanded] = useState(false);
-  const [isFromAddressCopied, setIsFromAddressCopied] = useState(false);
-  const [isToAddressCopied, setIsToAddressCopied] = useState(false);
-  const windowDimension = useWindowDimensions().width;
-
-  const onRawInputClick = () => {
-    setIsRawInputExpanded(!isRawInputExpanded);
-  };
-
-  return (
-    <div>
-      {/* desktop tablet */}
-      <div className="hidden md:block">
-        <div className="grid grid-rows-4 lg:grid-cols-6 md:grid-cols-4 lg:gap-x-0 gap-x-5">
-          <div>
-            <BoldedTitle
-              title={fixedTitle.gasDetail}
-              testId="desktop-transaction-gas-detail-title"
-            />
-          </div>
-          <div className="row-start-2 col-start-1">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="desktop-transaction-gas-price-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.gasPrice}
-              </div>
-              <Tooltip text={fixedTitle.gasPrice}>
-                <InfoIcon
-                  data-testid="desktop-transaction-gas-price-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-          </div>
-          <div className="row-start-2 col-start-2 lg:col-span-2">
-            <div className="flex">
-              <NumericFormat
-                data-testid="desktop-transaction-gas-price"
-                className="text-white-50 whitespace-normal tracking-[0.01em]"
-                thousandSeparator
-                value={gasPrice.value}
-                decimalScale={0}
-                suffix={` ${gasPrice.symbol}`}
-              />
-            </div>
-          </div>
-          <div className="row-start-3 col-start-1">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="desktop-transaction-gas-limit-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.gasLimit}
-              </div>
-              <Tooltip text={fixedTitle.gasLimit}>
-                <InfoIcon
-                  data-testid="desktop-transaction-gas-limit-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-          </div>
-          <div className="row-start-3 col-start-2 lg:col-span-2">
-            <div className="text-white-50 tracking-[0.01em] whitespace-normal">
-              <NumericFormat
-                data-testid="desktop-transaction-gas-limit"
-                thousandSeparator
-                value={gasLimit}
-                decimalScale={0}
-              />
-            </div>
-          </div>
-          <div className="row-start-4 col-start-1">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="desktop-transaction-gas-used-title"
-                className="text-white-700 w-[101px] tracking-[0.01em]"
-              >
-                {fixedTitle.gasUsed}
-              </div>
-              <Tooltip text={fixedTitle.gasUsed}>
-                <InfoIcon
-                  data-testid="desktop-transaction-gas-used-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-          </div>
-          <div className="row-start-4 col-start-2 lg:col-span-2">
-            <div className="flex flex-col">
-              <NumericFormat
-                data-testid="desktop-transaction-gas-used"
-                className="text-white-50 tracking-[0.01em] whitespace-normal"
-                thousandSeparator
-                value={gasUsed}
-                decimalScale={0}
-              />
-              <NumericFormat
-                data-testid="desktop-transaction-gas-used-percentage"
-                className="text-white-700 tracking-[0.02em] text-xs mt-1"
-                thousandSeparator
-                value={gasUsedPercentage}
-                decimalScale={2}
-                suffix="%"
-              />
-            </div>
-          </div>
-          {transactionType === fixedTitle.tokenized && (
-            <div className="lg:col-start-4 col-start-3 row-start-1 col-span-2">
-              <BoldedTitle
-                title={fixedTitle.tokenTransferred}
-                testId="desktop-transaction-token-transferred-title"
-              />
-            </div>
-          )}
-          {transactionType === fixedTitle.tokenized && (
-            <div className="lg:col-start-4 col-start-3 row-start-2">
-              <div
-                data-testid="desktop-transaction-token-transferred-from-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.from}
-              </div>
-            </div>
-          )}
-          {transactionType === fixedTitle.tokenized && (
-            <div className="lg:col-span-2 lg:col-start-5 col-start-4 row-start-2 ">
-              {isFromAddressCopied ? (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <LinkText
-                    testId="desktop-transaction-token-transferred-from-copied"
-                    label={fixedTitle.copied}
-                    href={`/address/${from}`}
-                  />
-                  <GreenTickIcon data-testid="desktop-transaction-token-transferred-from-green-tick-icon" />
-                </div>
-              ) : (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <LinkText
-                    customStyle="tracking-[0.01em]"
-                    testId="desktop-transaction-token-transferred-from"
-                    label={truncateTextFromMiddle(
-                      from,
-                      windowDimension <= 1280 ? 4 : 11
-                    )}
-                    href={`/address/${from}`}
-                  />
-                  <FiCopy
-                    role="button"
-                    data-testid="desktop-transaction-token-transferred-from-copy-icon"
-                    onClick={() =>
-                      onCopyAddressIconClick(setIsFromAddressCopied, from)
-                    }
-                    className="text-white-50"
-                  />
-                </div>
-              )}
-            </div>
-          )}
-          {transactionType === fixedTitle.tokenized && to && (
-            <div className="lg:col-start-4 col-start-3 row-start-3">
-              <div
-                data-testid="desktop-transaction-token-transferred-to-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.to}
-              </div>
-            </div>
-          )}
-          {transactionType === fixedTitle.tokenized && to && (
-            <div className="lg:col-start-5 col-start-4 row-start-3 lg:col-span-2">
-              {isToAddressCopied ? (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <LinkText
-                    testId="desktop-transaction-token-transferred-to-copied"
-                    label={fixedTitle.copied}
-                    href={`/address/${to}`}
-                  />
-                  <GreenTickIcon data-testid="desktop-transaction-token-transferred-to-green-tick-icon" />
-                </div>
-              ) : (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <LinkText
-                    testId="desktop-transaction-token-transferred-to"
-                    label={truncateTextFromMiddle(
-                      to,
-                      windowDimension <= 1280 ? 4 : 11
-                    )}
-                    href={`/address/${to}`}
-                  />
-                  <FiCopy
-                    role="button"
-                    data-testid="transaction-token-transferred-to-copy-icon"
-                    onClick={() =>
-                      onCopyAddressIconClick(setIsToAddressCopied, to)
-                    }
-                    className="text-white-50 h-[22px]"
-                  />
-                </div>
-              )}
-            </div>
-          )}
-          {transactionType === fixedTitle.tokenized && (
-            <div className="lg:col-start-4 col-start-3 row-start-4">
-              <div
-                data-testid="desktop-transaction-token-transferred-for-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.for}
-              </div>
-            </div>
-          )}
-
-          {/* Commented because we have no token txs for now
-           {type === fixedTitle.tokenized && (
-            <div className="lg:col-start-5 col-start-4 row-start-4 col-span-2">
-              <NumericFormat
-                data-testid="desktop-transaction-token-transferred-for"
-                className="text-white-50 tracking-[0.01em] whitespace-normal"
-                thousandSeparator
-                value={forToken.value}
-                decimalScale={8}
-                suffix={` ${forToken.symbol}`}
-              />
-            </div>
-          )} */}
-        </div>
-      </div>
-      {/* mobile */}
-      <div className="md:hidden flex flex-col gap-y-16">
-        <div className="flex flex-col gap-y-6">
-          <BoldedTitle
-            title={fixedTitle.gasDetail}
-            testId="transaction-gas-detail-title"
-          />
-          <div className="flex flex-row">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="transaction-gas-price-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.gasPrice}
-              </div>
-              <Tooltip text={fixedTitle.gasPrice}>
-                <InfoIcon
-                  data-testid="transaction-gas-price-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-            <div className="flex text-right">
-              <NumericFormat
-                data-testid="transaction-gas-price"
-                className="text-white-50 whitespace-normal tracking-[0.01em]"
-                thousandSeparator
-                value={gasPrice.value}
-                decimalScale={8}
-                suffix={` ${gasPrice.symbol}`}
-              />
-            </div>
-          </div>
-          <div className="flex flex-row">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="transaction-gas-limit-title"
-                className="text-white-700 tracking-[0.01em]"
-              >
-                {fixedTitle.gasLimit}
-              </div>
-              <Tooltip text={fixedTitle.gasLimit}>
-                <InfoIcon
-                  data-testid="transaction-gas-limit-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-            <div className="text-right">
-              <div className="text-white-50 tracking-[0.01em] whitespace-normal">
-                {useUnitSuffix(gasLimit)}
-              </div>
-            </div>
-          </div>
-          <div className="flex flex-row">
-            <div className="flex flex-row items-center grow">
-              <div
-                data-testid="transaction-gas-used-title"
-                className="text-white-700 w-[101px] tracking-[0.01em]"
-              >
-                {fixedTitle.gasUsed}
-              </div>
-              <Tooltip text={fixedTitle.gasUsed}>
-                <InfoIcon
-                  data-testid="transaction-gas-used-icon"
-                  className="ml-[9.33px] mr-[8.67px]"
-                />
-              </Tooltip>
-            </div>
-            <div className="flex flex-col">
-              <NumericFormat
-                data-testid="transaction-gas-used"
-                className="text-white-50 tracking-[0.01em] whitespace-normal text-right"
-                thousandSeparator
-                value={gasUsed}
-                decimalScale={0}
-              />
-              <NumericFormat
-                data-testid="transaction-gas-used-percentage"
-                className="text-white-700 tracking-[0.02em] text-xs mt-1 text-right"
-                thousandSeparator
-                value={gasUsedPercentage}
-                decimalScale={2}
-                suffix="%"
-              />
-            </div>
-          </div>
-        </div>
-        {transactionType === TransactionType.Tokenized && (
-          <div className="flex flex-col gap-y-6">
-            <BoldedTitle
-              title={fixedTitle.tokenTransferred}
-              testId="transaction-token-transferred-title"
-            />
-            <div className="flex flex-row">
-              <div className="grow">
-                <div
-                  data-testid="transaction-token-transferred-from-title"
-                  className="text-white-700 tracking-[0.01em]"
-                >
-                  {fixedTitle.from}
-                </div>
-              </div>
-              {isFromAddressCopied ? (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <LinkText
-                    testId="transaction-token-transferred-from-copied"
-                    label={fixedTitle.copied}
-                    href={`/address/${from}`}
-                  />
-                  <GreenTickIcon data-testid="transaction-token-transferred-from-green-tick-icon" />
-                </div>
-              ) : (
-                <div className="flex flex-row items-center gap-x-2.5">
-                  <div className={clsx({ "mr-[10px]": isFromAddressCopied })}>
-                    <LinkText
-                      customStyle="tracking-[0.01em]"
-                      testId="transaction-token-transferred-from"
-                      label={truncateTextFromMiddle(
-                        from,
-                        windowDimension <= 1280 ? 4 : 11
-                      )}
-                      href={`/address/${from}`}
-                    />
-                  </div>
-                  <FiCopy
-                    role="button"
-                    data-testid="transaction-token-transferred-from-copy-icon"
-                    onClick={() =>
-                      onCopyAddressIconClick(setIsFromAddressCopied, from)
-                    }
-                    className="text-white-50"
-                  />
-                </div>
-              )}
-            </div>
-            {to && (
-              <div className="flex flex-row">
-                <div className="grow">
-                  <div
-                    data-testid="transaction-token-transferred-to-title"
-                    className="text-white-700"
-                  >
-                    {fixedTitle.to}
-                  </div>
-                </div>
-                <div>
-                  {isToAddressCopied ? (
-                    <div className="flex flex-row items-center gap-x-2.5">
-                      <LinkText
-                        testId="transaction-token-transferred-to-copied"
-                        label={fixedTitle.copied}
-                        href={`/address/${to}`}
-                      />
-                      <GreenTickIcon data-testid="transaction-token-transferred-to-green-tick-icon" />
-                    </div>
-                  ) : (
-                    <div className="flex flex-row items-center gap-x-2.5 ">
-                      <div className={clsx({ "mr-[10px]": isToAddressCopied })}>
-                        <LinkText
-                          testId="transaction-token-transferred-to"
-                          label={truncateTextFromMiddle(
-                            to,
-                            windowDimension <= 1280 ? 4 : 11
-                          )}
-                          href={`/address/${to}`}
-                        />
-                      </div>
-                      <FiCopy
-                        role="button"
-                        data-testid="transaction-token-transferred-to-copy-icon"
-                        onClick={() =>
-                          onCopyAddressIconClick(setIsToAddressCopied, to)
-                        }
-                        className="text-white-50"
-                      />
-                    </div>
-                  )}
-                </div>
-              </div>
-            )}
-            {/* Commented because we have no token txs for now
-            <div className="flex flex-row">
-              <div className="grow">
-                <div
-                  data-testid="transaction-token-transferred-for-title"
-                  className="text-white-700 tracking-[0.01em]"
-                >
-                  {fixedTitle.for}
-                </div>
-              </div>
-              <div className="flex lg:text-left md:text-left text-right">
-                <NumericFormat
-                  data-testid="transaction-token-transferred-for"
-                  className="text-white-50 tracking-[0.01em]"
-                  thousandSeparator
-                  value={forToken.value}
-                  decimalScale={8}
-                  suffix={` ${forToken.symbol}`}
-                />
-              </div> 
-            </div> */}
-          </div>
-        )}
-      </div>
-      <div>
-        <BoldedTitle
-          className="lg:pt-14 md:pt-14 pt-16 pb-6"
-          testId="transaction-raw-input-title"
-          title={fixedTitle.rawInput}
-        />
-        <div
-          className={clsx("flex flex-row items-center lg:mb-[14px] mb-3", {
-            "mb-3": isRawInputExpanded,
-          })}
-        >
-          <div
-            data-testid="transaction-hex-title"
-            role="button"
-            tabIndex={0}
-            onKeyDown={onRawInputClick}
-            onClick={onRawInputClick}
-            className="text-white-50 pr-[10.29px]"
-          >
-            {fixedTitle.hex}
-          </div>
-          <div
-            data-testid="transaction-hex-icon"
-            role="button"
-            tabIndex={0}
-            onKeyDown={onRawInputClick}
-            onClick={onRawInputClick}
-          >
-            <div className="grid ">
-              <FiChevronDown
-                size={24}
-                className={clsx(
-                  "col-start-1 row-start-1 text-white-700",
-                  isRawInputExpanded
-                    ? "transition-opacity duration-300 ease-out opacity-0"
-                    : "opacity-100"
-                )}
-              />
-              <FiChevronUp
-                size={24}
-                className={clsx(
-                  "col-start-1 row-start-1 text-white-700",
-                  isRawInputExpanded
-                    ? "opacity-100"
-                    : "transition-opacity duration-300 ease-out opacity-0"
-                )}
-              />
-            </div>
-          </div>
-        </div>
-        {isRawInputExpanded && (
-          <div
-            data-testid="transaction-hex"
-            className={clsx(
-              " transition-opacity rounded-lg font-space-mono tracking-[-0.04em] break-all bg-black-800 lg:py-6 md:py-5 py-4 lg:px-10 md:px-8 px-4 border-[1px] border-black-600 text-white-50 text-xs"
-            )}
-          >
-            {hex}
           </div>
         )}
       </div>
